Mega changes for version 0.2.0 (read the CHANGELOG).

+++ fusefs/mount/mntopts.h
@@ -78,7 +78,7 @@
 struct mntopt {
 	const char *m_option;	/* option name */
 	int m_inverse;		/* if a negative option, eg "dev" */
-	int m_flag;		/* bit to set, eg. MNT_RDONLY */
+	uint64_t m_flag;	/* bit to set, eg. MNT_RDONLY */
 	int m_altloc;		/* 1 => set bit in altflags */
 };
 
@@ -124,7 +124,7 @@ struct mntopt {
         MOPT_AUTOMOUNTED,						\
 	MOPT_DEFWRITE
 	
-void getmntopts __P((const char *, const struct mntopt *, int *, int *));
+void getmntopts __P((const char *, const struct mntopt *, int *, uint64_t *));
 void checkpath __P((const char *, char resolved_path[]));
 void rmslashes __P((char *, char *));
 extern int getmnt_silent;
